Fixed contact icons opening
The contact icons now open correctly with a click on the icon.
The contact icons now close correctly with a click on the icon.
The nav bar was overlapping the icon-wrappersleading to  troubles to click the mail icon. Fixed now.
Added transparency to the scrolling button.
Changed the scrolling delays.
Changed to scrolling button rotating delay.